Remove pipe.listenFn Separate pipe states into state.go Merge message.go to pipe.go
